v0.1.1: 100-200x query speedup, security fixes, Helm chart
Performance:
- Table registry: keep tables open in memory for zero-overhead scans
- Global mmap cache: eliminate repeated mmap/munmap syscalls
- Metadata & partition list caching with invalidation on writes
- Skip optimizer for simple queries (no WHERE/ORDER BY)
- Limit pushdown into partition scanner (scan 1 partition, not all)
- Pre-warm query engine on server start
- Remove spawn_blocking for query execution
- Web console shows server-side query timing (not HTTP round-trip)
- Result: SELECT * LIMIT 25 from 2-4ms to 11µs (200x faster)

Security:
- Wire up validate_table_name on all write paths (path traversal fix)
- Service account hashing: SHA-256 -> Argon2id (with legacy compat)
- Auth error JSON injection fix (format! -> serde_json)
- Remove /api/v1/diagnostics from public paths
- ILP auth response size limit (4KB max)

CLI:
- config show/validate/generate commands
- check all/wal/partitions/metadata/disk-usage commands
- replication status/promote/demote commands
- debug wal-inspect/partition-info/column-dump/diagnostics commands
- compact (WAL merge) and status (remote health) commands
- version command with build info

Infrastructure:
- Helm chart for Kubernetes (StatefulSet, replication, monitoring)
- GitHub community files (LICENSE, CoC, SECURITY, CONTRIBUTING)
- Issue templates, PR template, CODEOWNERS, dependabot
- Reference config file (exchange-db.example.toml)
- Docs index, CLI reference, production checklist

Code quality:
- Zero compiler warnings across all 6 crates
- Zero clippy warnings (all lints resolved)
- 1744 tests passing, 0 failures

+++ crates/server/src/bench.rs
@@ -116,7 +116,7 @@ fn run_query(db_root: &Path, sql: &str) -> (Duration, u64) {
 
     let row_count = match &result {
         QueryResult::Rows { rows, .. } => rows.len() as u64,
-        QueryResult::Ok { affected_rows } => *affected_rows as u64,
+        QueryResult::Ok { affected_rows } => *affected_rows,
     };
 
     (elapsed, row_count)
